% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.defunct.r
\name{build.template}
\alias{build.template}
\title{Deprecated functions in geomorph}
\usage{
build.template()
}
\description{
The following functions are now deprecated
}
\details{
This function has been deprecated.
}
