\name{curves2d}
\alias{curves2d}
\title{Deprecated functions in geomorph}
\usage{
curves2d()
}
\description{
The following functions have been deprecated in geomorph
}
\details{
This function has been deprecated. Below shows the original
function and the replacement function.

curves2d: use define.sliders.2d

digit.curves: use define.sliders.3d
}

