% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.modules.r
\name{define.modules}
\alias{define.modules}
\title{Define modules (landmark partitions)}
\usage{
define.modules(spec, nmodules)
}
\arguments{
\item{spec}{A p x k matrix containing landmark coordinates of a single specimen (2D)}

\item{nmodules}{Number of modules to be defined}
}
\value{
Function returns a vector of which landmarks belong in which module (e.g. 1, 1, 1, 2, 2, 3, 3, 3, 2) to be used
with \code{\link{modularity.test}} or \code{\link{integration.test}}.
}
\description{
An interactive function to define which landmarks should be assigned to each module (landmark partition).
}
\details{
Function takes a matrix of digitized landmark coordinates (e.g. from \code{\link{mshape}}) and allows the user to assign 
landmarks to each module. The output is a list of which 
landmarks belong in which partition, to be used by \code{\link{modularity.test}} or \code{\link{integration.test}}.  
 
 \subsection{Selection in 2D}{ 
Choosing which landmarks will be included in each module involves landmark selection using a mouse in 
the plot window. The user is prompted to select each landmark in turn to be assigned to module 1: using the LEFT mouse button 
(or regular button for Mac users), click on the hollow circle to choose the landmark. Selected landmarks 
will be filled in. When all landmarks for module 1 are chosen, press 'esc', and then start selecting
landmarks for module 2. Repeat until all modules are defined.
}

 \subsection{Notes for geomorph 4.1}{ 
 Starting with geomorph version 4.1, interactive module selection in 3D is no longer
 supported, as RGL is not supported under MacOS Tahoe.
 }
}
\seealso{
\code{\link{modularity.test}} and \code{\link{integration.test}}
}
\author{
Dean Adams and Emma Sherratt
}
\keyword{utilities}
