% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\name{define.sliders.2d}
\alias{define.sliders.2d}
\title{Defunct functions in geomorph}
\usage{
define.sliders.2d()
}
\description{
The following functions are no longer supported in geomorph
}
\details{
define.sliders.2d: use define.sliders
}

