% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.defunct.r
\name{digit.fixed}
\alias{digit.fixed}
\title{This function has been deprecated, as RGL no longer}
\usage{
digit.fixed()
}
\description{
This function has been deprecated, as RGL no longer
}
