\name{editTemplate}
\alias{editTemplate}
\title{Edit 3D template}
\usage{
  editTemplate(template, fixed, n)
}
\arguments{
  \item{template}{Matrix of template 3D coordinates.}

  \item{fixed}{Number of "fixed" landmark points (non
  surface sliding points)}

  \item{n}{Number of points to be removed}
}
\description{
  A function to edit the 3D template file by removing
  undesirable points.
}
\author{
  Erik Otarola-Castillo
}
\keyword{editTemplate}

