% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\name{geomorph-package}
\alias{geomorph-package}
\alias{geomorph}
\title{Geometric morphometric analyses for 2D/3D data}
\description{
Functions in this package allow one to read and manipulate landmark data; generate shape
 variables via Procrustes analysis for points, curves and surface data, perform statistical analyses
 of shape variation and covariation, and provide graphical depictions of shapes and patterns of
 shape variation.
}
\section{geomorph TOC}{

geomorph-package
}

\author{
Dean C. Adams, Michael Collyer, Antigoni Kaliontzopoulou, and Erica Baken
}
