% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\docType{data}
\name{hummingbirds}
\alias{hummingbirds}
\title{Landmark data from hummingbird bills (includes sliding semilandmarks on curves)}
\description{
Landmark data from hummingbird bills (includes sliding semilandmarks on curves)
}
\author{
Chelsea Berns and Dean Adams
}
\references{
Berns, C.M., and Adams, D.C. 2010. Bill shape and sexual shape dimorphism between two species
of temperate hummingbirds: Archilochus alexandri (black-chinned hummingbirds) and Archilochus colubris
(ruby-throated hummingbirds). The Auk. 127:626-635.
}
\keyword{datasets}

