% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlmkdist.r
\name{interlmkdist}
\alias{interlmkdist}
\title{Calculate linear distances between landmarks}
\usage{
interlmkdist(A, dists)
}
\arguments{
\item{A}{A 3D array (p x k x n) containing landmark coordinates for a set of specimens}

\item{dists}{A matrix (m x 2) of landmark addresses for the start and end landmarks defining m linear measurements}
}
\value{
Function returns a matrix (n x m) of m linear distances for n specimens
}
\description{
A function to calculate linear distances between a set of landmark coordinates (interlandmark distances)
}
\details{
Function takes a 3D array of landmark coordinates from a set of specimens and a matrix 
of addresses for the start and end landmarks defining linear measurements and then
calculates the interlandmark distances. The function returns a matrix of linear distances for all specimens.
If the 'dists' matrix has row names defining the name of the linear measurements, the returned matrix will use
these for column names (see example).
}
\examples{
 
data(plethodon)
# Make a matrix defining three interlandmark distances 
dists <- matrix(c(8,9,6,12,4,2), ncol=2, byrow=T, dimnames = list(c("eyeW", "headL", "mouthL"),c("start", "end")))
# where 8-9 is eye width; 6-12 is head length; 4-2 is mouth length
A <- plethodon$land
lineardists <- interlmkdist(A, dists)
}
\author{
Emma Sherratt
}
\keyword{utilities}

