% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K.modules.r
\name{K.modules}
\alias{K.modules}
\title{K-modules simulation}
\usage{
K.modules(
  ME,
  rel.dims = NULL,
  min.lmk = 2,
  nsims = 1000,
  seed = NULL,
  print.progress = FALSE
)
}
\arguments{
\item{ME}{a module.eigen object, see \code{\link{module.eigen}}.}

\item{rel.dims}{The number of dimensions (eigenvectors) to sum for ranking outcomes.  If NULL, the relevant
dimensions from the module.eigen results will be used.  If more dimensions than ar epossible are chosen, the
number will be truncated.}

\item{min.lmk}{A numeric value to indicate the minimum number of landmarks (or variables) for a module.
If this value is larger than p/K, for p landmarks in K partitions, it will be adjusted to be the largest integer 
equal to or less than p/k.}

\item{nsims}{The number of simulations to use.}

\item{seed}{An optional value for seed control of simulations.  The value can be NULL, "random",
or an integer.  See  \code{\link{set.seed}}}

\item{print.progress}{A logical value to indicate whether a progress bar should be printed to the screen. 
This is helpful for long-running analyses.}

\item{K}{The number of modules into which the data are partitioned.
which landmarks (or variables) belong in which partition: 
(e.g. A, A, A, B, B, B, C, C, C). This is the same as partition.gp in, e.g., \code{\link{module.eigen}}.  If provided, it
will be one of simulated hypotheses considered.}
}
\value{
Objects of class "K.modules" return a list of the following:
 \item{eig.sums}{A rank-ordered vector of the sum of relevant eigenvalues.}
 \item{modules}{A rank-ordered list of modular hypotheses, consistent with eigenvalues.  Modules are 
 categories presented numerically, irrespective of factor levels used for an a priori modular hypothesis.}
 \item{hypothesis}{Whether an a priori hypothesis was used.}
 \item{hypothesis.rank}{If an a priori hypothesis was used, where it ranks among all simulated.}
 \item{mean}{If coordinate data are input, the mean configuration for use in plotting tools.}
 \item{VCV}{The (variance-) covariance matrix produced, for use in plotting tools.}
 \item{A}{The array of shape coordinates, if available, passed on for plotting purposes.}
 \item{eig1.ref}{The first eigenvalue or sum of eignvalues for relevant dimensions, from eigenanalysis of the total coavriance matrix.
 This value is used to relative the eigenvalue sums in random simulations.}
 \item{rel.dims}{The relevant dimensions used in the analysis.}
}
\description{
Function finds many possible permutations of K modules and ranks them in terms of 
greatest maximum eigenvalue or summed eigenvalues for relevant vectors,
based on covariance matrices for modules (see \code{\link{module.eigen}}).
}
\details{
The function simulates a desired number of modular hypotheses for a set of morphometric data, and 
ranks them in terms of either their first eigenvalue or sum of a number of eigenvalues, which can
be directed or assumed from a module eigen-analysis performed on a covariance matrix 
containing only modular covariances (see \code{\link{module.eigen}}).  An a priori modular hypothesis can 
be included (same input as in \code{\link{modularity.test}}, \code{\link{phylo.modularity}}, and 
\code{\link{module.eigen}}), 
and the minimum number of landmarks in a module can be adjusted.  A module.eigen object
 \code{\link{module.eigen}} is required for input. 

The function simulates many modular hypotheses and partitions covariance matrices by modules 
(see \code{\link{module.eigen}}),
performs eigen-analysis, and retains the eigenvalues along with the hypothesis.  All simulated 
outcomes (summed eigenvalues for relevant dimensions) are then 
rank-ordered to ascertain which modular hypotheses produce the strongest modularity.  Note that the 
analysis is "blind"
in terms of anatomical feasibility; thus, a good solution might identify modules comprising integrative 
rather than modular 
covariances.  Plotting tools found in \code{\link{plot.K.modules}} should help elucidate if this is the case.  
This function 
is probably only appealing if used with \code{\link{plot.K.modules}} and \code{\link{summary.K.modules}}.  
Examples below
should demonstrate how these tools can be used with real data.
}
\examples{

data(plethspecies)
Y.gpa<-gpagen(plethspecies$land)    #GPA-alignment
land.gps<-c("A","A","A","A","A","B","B","B","B","B","B") 

}
\references{
Collyer et al. In review.
}
\seealso{
\code{\link{summary.K.modules}}, \code{\link{plot.K.modules}}
\code{\link{module.eigen}}, \code{\link{modularity.test}}, 
\code{\link{phylo.integration}}, \code{\link{phylo.modularity}},
}
\author{
Michael Collyer
}
\keyword{analysis}
