% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\docType{data}
\name{larvalMorph}
\alias{larvalMorph}
\title{Head and tail shapes of larval salamanders}
\description{
Landmark data from heads and tails of larval Salamanders exposed to different treatments of herbicides.
}
\details{
Data set includes tail landmarks (coords), index for identifying semilandmarks (sliders), and vectors
for variables including herbicide treatment (Treatment) and Family (Family).  The latter variable indicates
the egg masses (clutches) sampled in the wild from which individual eggs were randomly assigned to treatment.
See Levis et al. (2016) for more experimental details.
}
\references{
Levis, N.A, M.L. Schooler, J.R. Johnson, and M.L. Collyer. 2016. The effects of terrestrial and aquatic herbicides on
larval salamander morphology and swim speed. Biological Journal of the Linnean Society.  Accepted.
}
\author{
Michael Collyer
}
\keyword{datasets}
