% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\docType{data}
\name{lizards}
\alias{lizards}
\title{Dorsal head shape data of lizards}
\description{
Superimposed landmark data of the dorsal view of lizard heads
}
\details{
Dataset includes superimposed landmarks (coords), centroid size (cs), an index of individuals (ind) and digitizing repetitions (rep), and a table of symmetrical matching
landmarks (lm.pairs). The object is a \code{\link{geomorph.data.frame}}.
The dataset corresponds to the data for population "b" from Lazic et al. 2015.
}
\references{
Lazic, M., Carretero, M.A., Crnobrnja-Isailovic, J. & Kaliontzopoulou, A. 2015. Effects of
environmental disturbance on phenotypic variation: an integrated assessment of canalization, developmental 
stability, modularity and allometry in lizard head shape. American Naturalist 185: 44-58.
}
\author{
Antigoni Kaliontzopoulou
}
\keyword{datasets}
