% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.map.r
\name{module.map}
\alias{module.map}
\title{Make a map of module covariances}
\usage{
module.map(
  A,
  partition.gp,
  mincol = "white",
  maxcol = "dark red",
  col0 = "white",
  bins = 10,
  near0 = TRUE,
  phy = NULL,
  Cov = NULL,
  transform. = TRUE,
  label.vars = FALSE,
  label.mods = TRUE,
  label.cex = 0.5,
  borders = TRUE
)
}
\arguments{
\item{A}{A 3D array (p x k x n) containing Procrustes shape variables for all specimens, or a matrix (n x variables)}

\item{partition.gp}{A vector for an a priori, modular hypothesis, explaining
which landmarks (or variables) belong in which partition: 
(e.g. A, A, A, B, B, B, C, C, C). This is the same as partition.gp in, e.g., \code{\link{module.eigen}}.}

\item{mincol}{The plot color for the minimum, non-0 covariance found.}

\item{maxcol}{The plot color for the maximum covariance found.}

\item{col0}{The plot color for 0 (lacking) covariance.}

\item{bins}{The number  of color bins to us for covariance gradient.}

\item{near0}{A logical value for whether covariances near 0 should be considered 0.  This is especially helpful
for geometric morphometric data that tend to be correlated because of generalized Procrustes analysis.  If TRUE,
more of the module map will be colored the same as col0, except the more intense covariances.}

\item{phy}{Optional argument to include a class \code{phylo} phylogenetic tree.  Tip labels must match data names.
This argument instructs the function to estimate a phylogenetic covariance matrix based on a Brownian motion model
of evolutionary divergence.  The Cov argument allows a user to define a hypothetical covariance matrix, if different 
than a BM model.}

\item{Cov}{Optional argument to include a hypothetical covariance matrix used for non-independence of observations.
Row and column names must match data names.  If both a phy and Cov are provided, Cov will override phy.}

\item{transform.}{A logical argument for whether to use transformed residuals, if a phylogeny is provided.  If TRUE,
a GLS covariance matrix will be estimated; if FALSE, data will be centered on GLS mean but an OLS covariance matrix 
will be estimated.  The former is more representative of covariances independent of phylogeny;  the latter 
is more representative of dispersion in the tangent space.  See \code{\link{module.eigen}}.}

\item{label.vars}{A logical variable for whether to include variable names as row and column labels.}

\item{label.mods}{A logical variable for whether to numerically label modules.}

\item{label.cex}{Magnification of the variable labels in the plot.}

\item{borders}{A logical variable for whether to add borders (lines) around modules.}
}
\description{
Function makes a plot of a covariance matrix arrange by modules, with colors that
measure the intensity of covariances within and between modules.
}
\examples{

# OLS vs GLS example

data(plethspecies)
Y.gpa<-gpagen(plethspecies$land)    #GPA-alignment
land.gps<-c("A","A","A","A","A","B","B","B","B","B","B") 

# OLS approach

module.map(Y.gpa$coords, partition.gp = land.gps)

# GLS-centered approach.  This approach find the GLS mean but does not
# transform residuals.  Thus, it is produces OLS-like covariance matrices
# that are centered on the GLS mean rather than the OLS mean

module.map(Y.gpa$coords, partition.gp = land.gps,
phy = plethspecies$phy, transform. = FALSE)

# GLS-transformed approach.  This approach find the GLS mean and
# transforms residuals.  Thus, it produces GLS covariance matrices,
# from residuals rendered independent of phylogenetic covariances.

module.map(Y.gpa$coords, partition.gp = land.gps,
phy = plethspecies$phy, transform. = TRUE)

# Multi-module-hypothesis example
data(pupfish)

# Two-module hypothesis
hyp2 <- c(rep(1, 3), 2, rep(1, 5),
rep(2, 9), rep(1, 20), rep(2, 18))

# Three-module hypothesis
hyp3 <- c(rep(1, 3), 2, 1, rep(3, 4),
rep(2, 9), rep(1, 12), rep(3, 8), rep(2, 18))

par(mfrow = c(1, 2))
plot(pupfish$coords[,,1], pch = 21, bg = hyp2, asp = 1)
plot(pupfish$coords[,,1], pch = 21, bg = hyp3, asp = 1)

module.map(pupfish$coords, partition.gp = hyp2)
module.map(pupfish$coords, partition.gp = hyp3)

par(mfrow = c(1, 1))
}
\references{
Collyer et al. In review.
}
\seealso{
\code{\link{module.eigen}}, 
\code{\link{summary.module.eigen}}, \code{\link{plot.module.eigen}}
\code{\link{two.b.pls}}, \code{\link{modularity.test}}, 
\code{\link{phylo.integration}}, \code{\link{phylo.modularity}},
and \code{\link{compare.pls}}
}
\author{
Michael Collyer
}
\keyword{analysis}
