% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\docType{data}
\name{motionpaths}
\alias{motionpaths}
\title{Simulated motion paths}
\description{
Simulated motion paths
}
\author{
Dean Adams
}
\references{
Adams, D. C., and M. L. Collyer. 2009. A general framework for the analysis of phenotypic
  trajectories in evolutionary studies. Evolution 63:1143-1154.
}
\keyword{datasets}

