% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\name{pairwiseD.test}
\alias{pairwiseD.test}
\title{Defunct functions in geomorph}
\usage{
pairwiseD.test()
}
\description{
The following functions are no longer supported in geomorph
}
\details{
pairwiseD.test: use advanced.procD.lm

pairwise.slope.test: use advanced.procD.lm
}

