% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/physignal.r
\name{physignal}
\alias{physignal}
\title{Assessing phylogenetic signal in morphometric data}
\usage{
physignal(A, phy, iter = 999)
}
\arguments{
\item{A}{A matrix (n x [p x k]) or 3D array (p x k x n) containing GPA-aligned coordinates for a set of specimens}

\item{phy}{A phylogenetic tree of {class phylo} - see \code{\link[ape]{read.tree}} in library ape}

\item{iter}{Number of iterations for significance testing}
}
\value{
Function returns a list with the following components:
  \item{phy.signal}{The estimate of phylogenetic signal}
  \item{pvalue}{The significance level of the observed signal}
  \item{random.K}{Each random K-statistic from random permutations}
  \item{permutations}{The number of random permutations used in the resampling procedure}
  \item{call}{The matched call}
}
\description{
Function calculates the degree of phylogenetic signal from a set of Procrustes-aligned specimens
}
\details{
The function estimates the degree of phylogenetic signal present in shape data for a given phylogeny.
It is assumed that the landmarks have previously been aligned
  using Generalized Procrustes Analysis (GPA) [e.g., with \code{\link{gpagen}}].
  The degree of phylogenetic signal in data is estimated using the multivariate version of the K-statistic
  (Kmult: Adams 2014). This value evaluates the degree of phylogenetic signal
  in a dataset relative to what is expected under a Brownian motion model of evolution. For geometric
  morphometric data, the approach is a mathematical generalization of the Kappa statistic (Blomberg et al.
  2003) appropriate for highly multivariate data (see Adams 2014).Significance testing
  is found by permuting the shape data among the tips of the phylogeny. Note that this
  method can be quite slow as ancestral states must be estimated for every iteration.

This function can also be used with univariate data (i.e. centroid size) if imported as matrix with rownames
giving the taxa names. In this case, the estimate of phylogenetic signal is identical to that found using the
standard kappa statistic (Blomberg et al. 2003).

The generic functions, \code{\link{print}}, \code{\link{summary}}, and \code{\link{plot}} all work with \code{\link{physignal}}.
 The generic function, \code{\link{plot}}, produces a histogram of random K statistics, associated with the resampling procedure.
}
\examples{
data(plethspecies)
Y.gpa<-gpagen(plethspecies$land)    #GPA-alignment

#Test for phylogenetic signal in shape
PS.shape <- physignal(A=Y.gpa$coords,phy=plethspecies$phy,iter=999)
summary(PS.shape)
plot(PS.shape)

#Test for phylogenetic signal in size
PS.size <- physignal(A=Y.gpa$Csize,phy=plethspecies$phy,iter=999)
summary(PS.size)
plot(PS.size)
}
\author{
Dean Adams
}
\references{
Blomberg SP, Garland T, Ives AR. 2003. Testing for phylogenetic signal in comparative
data: behavioral traits are more labile. Evolution, 57:717-745.

Adams, D.C. 2014. A generalized K statistic for estimating phylogenetic signal from shape and
other high-dimensional multivariate data. Systematic Biology.  63:685-697.
}
\keyword{analysis}

