\docType{data}
\name{plethland}
\alias{plethland}
\title{Landmark data from dataset plethodon}
\description{
  Landmark data from dataset plethodon
}
\author{
  Dean Adams
}
\references{
  Adams, D. C. 2004. Character displacement via aggressive
  interference in Appalachian salamanders. Ecology.
  85:2664-2670.

  Adams, D.C. 2010. Parallel evolution of character
  displacement driven by competitive selection in
  terrestrial salamanders. BMC Evolutionary Biology.
  10(72)1-10.
}
\seealso{
  \code{\link{plethodon}}
}
\keyword{data}
\keyword{datasets}

