% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\docType{data}
\name{plethodon}
\alias{plethodon}
\title{Landmark data from Plethodon salamander heads}
\description{
Landmark data from Plethodon salamander heads
}
\author{
Dean Adams
}
\references{
Adams, D. C. 2004. Character displacement via aggressive interference in Appalachian salamanders. 
Ecology. 85:2664-2670.

Adams, D.C. 2010. Parallel evolution of character displacement driven by competitive selection 
in terrestrial salamanders. BMC Evolutionary Biology. 10(72)1-10.
}
\keyword{datasets}

