% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\docType{data}
\name{plethShapeFood}
\alias{plethShapeFood}
\title{Head shape and food use data from Plethodon salamanders}
\description{
Head shape and food use data from Plethodon salamanders
}
\author{
Dean Adams
}
\references{
Adams, D. C., and F. J. Rohlf. 2000. Ecological character 
displacement in Plethodon: biomechanical differences found from a geometric 
morphometric study. Proceedings of the National Academy of Sciences, 
U.S.A. 97:4106-4111
}
\keyword{datasets}

