% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\docType{data}
\name{plethspecies}
\alias{plethspecies}
\title{Head shape and phylogenetic relationships for several Plethodon salamander species}
\description{
Head shape and phylogenetic relationships for several Plethodon salamander species
}
\references{
Phylogeny pruned from: Wiens et al. (2006). Evol.

Data from: Adams and Rohlf (2000); Adams et al. (2007); Arif et al. (2007) Myers and Adams (2008)
}
\author{
Dean Adams
}
\keyword{datasets}
