% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geomoprh.utils.R
\name{plot.CR.phylo}
\alias{plot.CR.phylo}
\title{Plot Function for geomorph}
\usage{
\method{plot}{CR.phylo}(x, ...)
}
\arguments{
\item{x}{plot object}

\item{...}{other arguments passed to plot}
}
\description{
Plot Function for geomorph
}
\author{
Dean Adams
}

