% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geomoprh.utils.R
\name{plot.gpagen}
\alias{plot.gpagen}
\title{Plot Function for geomorph}
\usage{
\method{plot}{gpagen}(x, ...)
}
\arguments{
\item{x}{plot object}

\item{...}{other arguments passed to plot}
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer
}

