% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{plot.K.modules}
\alias{plot.K.modules}
\title{Plot function for geomorph}
\usage{
\method{plot}{K.modules}(
  x,
  modules = 1:6,
  type = c("profile", "config", "pcoa"),
  relativize = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{K.modules}}}

\item{modules}{A vector or integer for which modules to consider in plot (should generally be small in length
if results are to be tractable).  If a profile plot is chosen, all eigenvalues will be plotted.  (The default value
is the highest ranking 6 modules.)}

\item{type}{One of "profile", config", or "pcoa" for plotting a profile of eigenvalues, 
landmark configurations with modules formatted differently, 
or a principal coordinate plot, respectively.}

\item{relativize}{A logical argument for whether to make eigenvalues a fraction of the maximum possible
(for profile plots).}

\item{...}{other arguments passed to plot, especially for changing points in a configuration.

This function provides plotting support for \code{\link{K.modules}}.  One of two plots can be produced: (1) a 
canvas of landmark configuration with points formatted differently for modules or (2) a principal coordinate plot, based
on Riemannian distances among modular covariance matrices.  Based on the modules requested (output from a K.modules
object), several configuration plots can be generated or one single principal coordinate plot with points equal in 
number to the vector length for "modules".  The length of this vector should be reasonable to avoid long computation of 
Riemannian distance among modular covariance matrices, or to avoid generating many configurations.  See 
\code{\link{K.modules}} for plotting examples.}
}
\description{
Plot function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
