% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{plot.mshape}
\alias{plot.mshape}
\title{Plot Function for geomorph}
\usage{
\method{plot}{mshape}(x, links = NULL, ...)
}
\arguments{
\item{x}{plot object (from \code{\link{mshape}})}

\item{links}{An optional matrix defining for links between landmarks}

\item{...}{other arguments passed to plot}
}
\description{
Plot Function for geomorph
}
\seealso{
\code{\link{define.links}}
}
\author{
Dean Adams and Antigoni Kaliontzopoulou
}
\keyword{utilities}
\keyword{visualization}
