% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{plot.physignal.eigen}
\alias{plot.physignal.eigen}
\title{Plot Function for geomorph}
\usage{
\method{plot}{physignal.eigen}(x, type = c("conf", "vectors", "both"), confidence = 0.95, ...)
}
\arguments{
\item{x}{plot object (from \code{\link{physignal.eigen}})}

\item{type}{the manner in which the distribution of permuted outcomes is represented:
confidence bands (type = "conf"), vectors (type = "vectors"), or both (type = "both")}

\item{...}{other arguments passed to plot}
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer and Dean Adams
}
\keyword{utilities}
\keyword{visualization}
