% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{plot.procD.allometry}
\alias{plot.procD.allometry}
\title{Plot Function for geomorph}
\usage{
\method{plot}{procD.allometry}(x, method = c("CAC", "RegScore", "PredLine"),
  warpgrids = TRUE, label = NULL, pt.col = NULL, mesh = NULL,
  shapes = TRUE, ...)
}
\arguments{
\item{x}{plot object}

\item{method}{Method for estimating allometric shape components; see \code{\link{procD.allometry}} for details}

\item{warpgrids}{A logical value indicating whether deformation grids for small and large shapes 
should be displayed (note: if groups are provided no TPS grids are shown)}

\item{label}{An optional vector indicating labels for each specimen that are to be displayed}

\item{pt.col}{An optional vector of colours to use for points (as in points(bg=))}

\item{mesh}{A mesh3d object to be warped to represent shape deformation of the minimum and maximum size 
if {warpgrids=TRUE} (see \code{\link{warpRefMesh}}).}

\item{shapes}{Logical argument whether to return the the shape coordinates shape coordinates of the small and large shapes}

\item{...}{other arguments passed to plot}
}
\value{
If shapes = TRUE, function returns a list containing the shape coordinates of the small and large shapes
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}

