% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.time.gm.prcomp.R
\name{plot.time.gm.prcomp}
\alias{plot.time.gm.prcomp}
\title{Create a 3d phylomorphospace plot of a \code{\link{gm.prcomp}} object with a phylogeny associated to it,
with time displayed on the z-axis.}
\usage{
\method{plot}{time.gm.prcomp}(
  x,
  axis1 = 1,
  axis2 = 2,
  tip.labels = TRUE,
  node.labels = TRUE,
  anc.states = TRUE,
  plot.param = list(tip.bg = "black", tip.cex = 1, node.bg = "grey", node.cex = 1,
    edge.color = "black", edge.width = 1, txt.cex = 1, txt.col = "black", txt.adj =
    c(-0.1, -0.1), node.txt.cex = 1, node.txt.col = "grey", node.txt.adj = c(-0.1, -0.1))
)
}
\arguments{
\item{x}{A \code{\link{gm.prcomp}} object}

\item{axis1}{A numeric value indicating which PC axis should be displayed as the X-axis}

\item{axis2}{A numeric value indicating which PC axis should be displayed as the Y-axis}

\item{tip.labels}{A logical value indicating whether taxa labels (tips) should be included}

\item{node.labels}{A logical value indicating whether node labels (ancestors) should be included}

\item{anc.states}{A logical value indicating whether points for the nodes should be included}

\item{plot.param}{A list of plotting parameters for the tips (tip.bg, tip.cex), nodes (node.bg, node.cex),
edges (edge.colour, edge.width), and labels of the tips (txt.cex, txt.col, txt.adj) and nodes 
(node.txt.cex, node.txt.col, node.txt.adj)}
}
\description{
This function replaces the 3d plotting functionality of plotGMPhyloMorphoSpace, now deprecated.
}
\examples{
data(plethspecies) 
Y.gpa<-gpagen(plethspecies$land)    #GPA-alignment    

PCA.w.phylo <- gm.prcomp(Y.gpa$coords, phy = plethspecies$phy)

plot.time.gm.prcomp(x)
plot.time.gm.prcomp(x, node.labels = F, anc.states = F, 
plot.param = list(tip.bg = "red", tip.cex = 3, 
edge.colour = "blue", edge.width = 2, txt.cex = 2))
}
\seealso{
\code{\link{gm.prcomp}}
}
\author{
Antigoni Kaliontzopoulou
}
\keyword{visualization}
