% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.R
\name{plot.trajectory.analysis}
\alias{plot.trajectory.analysis}
\title{Plot Function for geomorph}
\usage{
\method{plot}{trajectory.analysis}(x, ...)
}
\arguments{
\item{x}{plot object}

\item{...}{other arguments passed to plot}

\item{group.cols}{An optional vector of colors for group levels}
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer
}

