% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{print.bilat.symmetry}
\alias{print.bilat.symmetry}
\title{Print/Summary Function for geomorph}
\usage{
\method{print}{bilat.symmetry}(x, ...)
}
\arguments{
\item{x}{print/summary object (from \code{\link{bilat.symmetry}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
