% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{print.combined.set}
\alias{print.combined.set}
\title{Print/Summary Function for geomorph}
\usage{
\method{print}{combined.set}(x, ...)
}
\arguments{
\item{x}{print/summary object}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
