% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{print.compare.ZVrel}
\alias{print.compare.ZVrel}
\title{Print/Summary function for geomorph}
\usage{
\method{print}{compare.ZVrel}(x, ...)
}
\arguments{
\item{x}{print/summary object}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary function for geomorph
}
\author{
Dean Adams
}
\keyword{utilities}
