% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{print.geomorphShapes}
\alias{print.geomorphShapes}
\title{Print/Summary function for geomorph}
\usage{
\method{print}{geomorphShapes}(x, ...)
}
\arguments{
\item{x}{print/summary object}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
