% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{print.gm.prcomp}
\alias{print.gm.prcomp}
\title{Print/Summary function for geomorph}
\usage{
\method{print}{gm.prcomp}(x, ...)
}
\arguments{
\item{x}{print/summary object}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary function for geomorph
}
\author{
Antigoni Kaliontzopoulou
}
\keyword{utilities}
