% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{print.K.modules}
\alias{print.K.modules}
\title{Print/Summary function for geomorph}
\usage{
\method{print}{K.modules}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{K.modules}}}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
