% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{print.module.eigen}
\alias{print.module.eigen}
\title{Print/Summary function for geomorph}
\usage{
\method{print}{module.eigen}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{module.eigen}}}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
