% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{print.physignal.eigen}
\alias{print.physignal.eigen}
\title{Print/Summary Function for geomorph}
\usage{
\method{print}{physignal.eigen}(x, ...)
}
\arguments{
\item{x}{print/summary object (from \code{\link{physignal.eigen}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Dean Adams
}
\keyword{utilities}
