% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geomoprh.utils.R
\name{print.pls}
\alias{print.pls}
\title{Print/Summary Function for geomorph}
\usage{
\method{print}{pls}(x, ...)
}
\arguments{
\item{x}{print/summary object}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Michael Collyer
}

