% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{print.procD.allometry}
\alias{print.procD.allometry}
\title{Print/Summary Function for geomorph}
\usage{
\method{print}{procD.allometry}(x, ...)
}
\arguments{
\item{x}{print/summary object (from \code{\link{procD.allometry}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
