% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.defunct.r
\name{procD.allometry}
\alias{procD.allometry}
\title{Deprecated functions in geomorph}
\usage{
procD.allometry()
}
\description{
The following function has been deprecated in geomorph
}
\details{
This function has been deprecated. Use procD.lm, #' advanced.procD.lm now deprecated: use \code{\link{procD.lm}} instead, along with \code{\link{anova.lm.rrpp}} and \code{\link{pairwise}}.  
See examples in \code{\link{procD.lm}} 
and/or the vignette, 'Addition by subtraction: improving geomorph capabilities with fewer functions.'
}
