% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\docType{data}
\name{pupfish.ws}
\alias{pupfish.ws}
\title{Landmarks on pupfish}
\description{
Landmark data from pupfish body shapes, with indication of Sex, Poplulation and
Species from which fish were sampled. A phylogeny and phylogenetic covariance matrix are
included.
}
\details{
These data were previously aligned
with GPA.  Centroid size (CS) is also provided.
}
\references{
dams, D.C and M.L Collyer. 2024. Extending phylogenetic regression models for 
comparing within-species patterns across the tree of life. 
Methods in Ecology and Evolution. DOI: 10.1111/2041-210X.14438
}
\author{
Dean Adams and Michael Collyer
}
\keyword{datasets}
