\docType{data}
\name{rats}
\alias{rats}
\title{Landmark data from rat calvaria}
\description{
  Landmark data from rat calvaria
}
\author{
  Dean Adams
}
\references{
  Bookstein, F. L. 1991. Morphometric tools for landmark
  data: Geometry and Biology.  Cambridge Univ. Press, New
  York.
}
\keyword{data}
\keyword{datasets}

