% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ply.r
\name{read.ascii.ply}
\alias{read.ascii.ply}
\title{Internal function for reading PLY data in ASCII format}
\usage{
read.ascii.ply(plyfile)
}
\description{
Internal function for reading PLY data in ASCII format
}
