% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readland.fcsv.r
\name{readland.fcsv}
\alias{readland.fcsv}
\title{Read landmark data matrix from fcsv file}
\usage{
readland.fcsv(file = NULL, header_no = 3)
}
\arguments{
\item{file}{The name of a *.fcsv file containing three-dimensional landmark data to be read in}

\item{header_no}{The number of header lines in the fscv file, needed for knowing these lines are not data.}
}
\value{
Function returns a p x 3 matrix of x, y, z coordinates for p landmarks.

Note: to read in multiple files the following is useful:

filelist <- list.files(path = "PATH TO FOLDER with FILES", pattern = "*.fcsv", 
             full.names = TRUE)
mydata <- simplify2array(lapply(fcsv_files, readland.fcsv))
}
\description{
Read landmark data for a single specimen from an fcsv file obtained from SlicerMorph.
}
\details{
This function merely extracts x, y, z coordinates from an fcsv file.
}
\author{
Murat Maga, Michael Collyer and Dean Adams
}
\keyword{IO}
