% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\docType{data}
\name{scallops}
\alias{scallops}
\title{Landmark data from scallop shells}
\description{
Landmark data from scallop shells
}
\references{
Serb et al. (2011). "Morphological convergence of shell shape in distantly related
scallop species (Mollusca: Pectinidae)." Zoological Journal of the Linnean Society 163: 571-584.
}
\author{
Dean Adams and Erik Otarola-Castillo
}
\keyword{datasets}
