\docType{data}
\name{Specimen4Raw}
\alias{Specimen4Raw}
\title{Raw 3D scan data from scallop shell}
\description{
  Raw 3D scan data from scallop shell
}
\author{
  Dean Adams and Erik Otarola-Castillo
}
\references{
  Serb et al. (2011). "Morphological convergence of shell
  shape in distantly related scallop species (Mollusca:
  Pectinidae)." Zoological Journal of the Linnean Society
  163: 571-584.
}
\keyword{data}
\keyword{datasets}

