% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.R
\name{summary.advanced.procD.lm}
\alias{summary.advanced.procD.lm}
\title{Print/Summary Function for geomorph}
\usage{
\method{summary}{advanced.procD.lm}(object, ...)
}
\arguments{
\item{object}{print/summary object}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}

