% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geomoprh.utils.R
\name{summary.CR.phylo}
\alias{summary.CR.phylo}
\title{Print/Summary Function for geomorph}
\usage{
\method{summary}{CR.phylo}(object, ...)
}
\arguments{
\item{object}{print/summary object}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Dean Adams
}

