% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{summary.K.modules}
\alias{summary.K.modules}
\title{Print/Summary function for geomorph}
\usage{
\method{summary}{K.modules}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{K.modules}}}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
