% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{summary.physignal.z}
\alias{summary.physignal.z}
\title{Print/Summary Function for geomorph}
\usage{
\method{summary}{physignal.z}(object, ...)
}
\arguments{
\item{object}{print/summary object (from \code{\link{physignal.z}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
