% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/writeland.tps.r
\name{writeland.tps}
\alias{writeland.tps}
\title{Write landmark data to tps file}
\usage{
writeland.tps(A, file)
}
\arguments{
\item{A}{An array (p x k x n) containing landmark coordinates for a set of specimens}

\item{file}{Name of the *.tps file to be created}
}
\description{
Write *.tps file from obtain landmark coordinates in a 3-dimensional array
}
\details{
This function writes a *.tps file from a 3-dimensional array (p x k x n)
 of landmark coordinates.
}
\author{
Dean Adams
}
\keyword{IO}

