% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.defunct.r
\name{advanced.procD.lm}
\alias{advanced.procD.lm}
\title{Deprecated functions in geomorph}
\usage{
advanced.procD.lm()
}
\description{
The following function has been deprecated in geomorph
}
\details{
This function has been deprecated. Use \code{\link{procD.lm}} instead, along with \code{\link{anova.lm.rrpp}} and \code{\link{pairwise}}.  
See examples in \code{\link{procD.lm}} 
and/or the vignette, 'geomorph version 3.1.0 Assistance.'
}
