% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.defunct.r
\name{compare.modular.partitions}
\alias{compare.modular.partitions}
\title{Deprecated functions in geomorph}
\usage{
compare.modular.partitions()
}
\description{
The following function has been deprecated in geomorph
}
\details{
This function has been deprecated. Below shows the original function and the replacement function.

compare.modular.partitions now deprecated: use \code{\link{modularity.test}} instead
}

