\name{curves2d}
\alias{curves2d}
\title{Defunct functions in geomorph}
\usage{
curves2d()
}
\description{
The following functions are no longer supported in geomorph
}
\details{
This function is now defunct. Below shows the original
function and the replacement function.

curves2d: use define.sliders.2d

digit.curves: use define.sliders.3d
}

