% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/define.sliders.2d.r
\name{define.sliders.2d}
\alias{define.sliders.2d}
\title{Deprecated functions in geomorph}
\usage{
define.sliders.2d()
}
\description{
The following functions have been deprecated in geomorph
}
\details{
This function has been deprecated. Below shows the original function and the replacement function.

define.sliders.2d: use define.sliders

define.sliders.3d: use define.sliders
}

