% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/define.sliders.r
\name{define.sliders}
\alias{define.sliders}
\title{Select points to "slide" along curves.}
\usage{
define.sliders(spec, nsliders, surfsliders = FALSE)
}
\arguments{
\item{spec}{Name of specimen, as an object matrix containing 2D or 3D landmark coordinates}

\item{nsliders}{Number of landmarks to be semilandmarks that slide along curves}

\item{surfsliders}{Logical (3D only) If 'spec' contains landmarks that are "surface sliders",
made by \code{\link{buildtemplate}}, "surfslide.csv" should be in working directory}
}
\value{
Function returns a 'nsliders-x-3' matrix containing the landmark address of the curve sliders, indicating the
landmarks between which the slider landmarks will "slide". The matrix is also written to working directory
as "curveslide.csv". Matrix (or "curveslide.csv") is designed for use by \code{\link{gpagen}} during GPA.
}
\description{
An interactive function to define which landmarks will "slide" along two-dimensional (2D) or
three-dimensional (3D) curves.
}
\details{
Function takes a matrix of digitized landmark coordinates, such as made by \code{\link{digitize2d}} or \code{\link{digit.fixed}},
 and helps user choose which landmarks will be treated as "sliders" in Generalized Procrustes analysis
 \code{\link{gpagen}}. This type of semilandmark "slides" along curves lacking known landmarks
 (see Bookstein 1997 for algorithm details).
 Each sliding semilandmark ("sliders") will slide between two designated points, along a line
 tangent to the specified curvature.

Defining landmarks is an interactive procedure (see below for 2D and 3D routines). The procedure is overlapping.
For example: there are 5 landmarks (1:5), 1 and 5 are landmarks and 2,3,4 are sliders,
the user must select '1' '2' '3', and then '2' '3' '4', and then '3' '4' '5'.

\subsection{Selection in 2D}{
Choosing which landmarks will be sliders involves landmark selection using a mouse in the plot window.
To define the sliders, for each sliding landmark along the curve in the format 'before-slider-after',
using the LEFT mouse button (or regular button for Mac users), click on the hollow circle to choose the landmark
in the following order:
\enumerate{
 \item Click to choose the first landmark between which semi-landmark will "slide",
 \item Click to choose sliding landmark,
 \item Click to choose the last landmark between which semi-landmark will "slide",
 Selected landmarks will be filled in and lines are drawn connecting the three landmarks,
 and will highlight the sliding semilandmark in red and the flanking landmarks in blue.
}
}

#' \subsection{Selection in 3D}{
Choosing which landmarks will be sliders involves landmark selection using a mouse in the rgl plot window.
With a standard 3-button (PC) buildtemplate uses:
\enumerate{
 \item the RIGHT mouse button (primary) to choose points to be defined as sliders,
 \item the LEFT mouse button (secondary) is used to rotate mesh,
 \item the mouse SCROLLER (third/middle) is used to zoom in and out.
}
NOTE: Digitizing functions on MACINTOSH computers using a standard 3-button mice works as specified. Macs using platform
specific single button mice, XQuartz must be configured: go to Preferences > Input > tick "Emulate three button mouse":
\enumerate{
 \item press button to rotate 3D mesh,
 \item press button while pressing COMMAND key to select points to be defined as sliders,
 \item press button while pressing OPTION key to adjust mesh perspective.
 \item the mouse SCROLLER or trackpad two finger scroll is used to zoom in an out.
 }

To define the sliders, for each sliding landmark along the curve in the format 'before-slider-after':
\enumerate{
 \item Click to choose the first landmark between which semi-landmark will "slide",
 \item Click to choose sliding landmark,
 \item Click to choose the last landmark between which semi-landmark will "slide",
Screen will show lines connecting the three landmarks, and will highlight the sliding semilandmark in red.
}
}
}
\author{
Dean Adams, Erik Otarola-Castillo, Emma Sherratt
}
\references{
Bookstein, F. J. 1997 Landmark Methods for Forms without Landmarks: Morphometrics of
Group Differences in Outline Shape. Medical Image Analysis 1(3):225-243.
}
\seealso{
\code{\link{digitize2d}}, \code{\link{gpagen}}
}
\keyword{utilities}

