\name{digit.curves}
\alias{digit.curves}
\title{Deprecated functions in geomorph}
\usage{
digit.curves()
}
\description{
The following functions have been deprecated in geomorph
}
\details{
This function has been deprecated. Below shows the original
function and the replacement function.

curves2d: use define.sliders.2d

digit.curves: use define.sliders.3d
}

