% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTemplate.r
\name{editTemplate}
\alias{editTemplate}
\title{Edit 3D template}
\usage{
editTemplate(template, fixed, n)
}
\arguments{
\item{template}{Matrix of template 3D coordinates}

\item{fixed}{Number of "fixed" landmark points (non surface sliding points)}

\item{n}{Number of points to be removed}
}
\value{
Function returns a matrix containing the x,y,z coordinates of the new template landmarks.
Function also writes to the working directory 'template.txt' containing the x,y,z coordinates of the updated template
}
\description{
An interactive function to remove landmarks from a 3D template file.
}
\details{
The function edits a 'template.txt' file made by \code{\link{buildtemplate}}, which must be in the current
working directory, and which is overwritten. Use read.table("template.txt", header = T) to read in the template
first.
\subsection{Selection}{ 
Choosing which landmarks will be deleted involves landmark selection using a mouse in the rgl plot window. 
With a standard 3-button (PC) buildtemplate uses:
\enumerate{
 \item the RIGHT mouse button (primary) to choose points to be deleted,
 \item the LEFT mouse button (secondary) is used to rotate mesh, 
 \item the mouse SCROLLER (third/middle) is used to zoom in and out.
}
NOTE: Digitizing functions on MACINTOSH computers using a standard 3-button mice works as specified. Macs using platform 
specific single button mice, XQuartz must be configured: go to Preferences > Input > tick "Emulate three button mouse":
\enumerate{
 \item press button to rotate 3D mesh,
 \item press button while pressing COMMAND key to select points to be deleted,
 \item press button while pressing OPTION key to adjust mesh perspective.
 \item the mouse SCROLLER or trackpad two finger scroll is used to zoom in an out.
 }
}
}
\seealso{
\code{\link[rgl]{rgl-package}} (used in 3D plotting)
}
\author{
Erik Otarola-Castillo & Emma Sherratt
}
\keyword{digitizing}
