% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\docType{package}
\name{geomorph-package}
\alias{geomorph}
\alias{geomorph-package}
\title{Geometric morphometric analyses for 2D/3D data}
\description{
Geometric morphometric analyses for 2D/3D data
}
\author{
Dean C. Adams, Michael Collyer, Erik Otarola-Castillo, & Emma Sherratt

Functions in this package allow one to read, manipulate, and digitize landmark data; generate shape
 variables via Procrustes analysis for points, curves and surface data, perform statistical analyses
 of shape variation and covariation, and provide graphical depictions of shapes and patterns of
 shape variation.
}

