% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.prcomp.R
\name{gm.prcomp}
\alias{gm.prcomp}
\title{Principal components analysis of shape data}
\usage{
gm.prcomp(A, phy = NULL, phylo.pca = FALSE, Cov = NULL, ...)
}
\arguments{
\item{A}{A 3D array (p x k x n) containing landmark coordinates for a set of aligned specimens}

\item{phy}{An optional phylogenetic tree of class phylo - see \code{\link{read.tree}} in library ape}

\item{phylo.pca}{A logical value indicating whether phylogenetic PCA (TRUE) 
or the phylomorphospace approach (FALSE) should be used. See also details.}

\item{Cov}{An optional covariance matrix for weighting. See also details.}

\item{...}{Arguments passed on to \code{\link{prcomp}}.  By default, \code{\link{gm.prcomp}}
will attempt to remove redundant axes (eigenvalues effectively 0).  To override this, adjust the 
argument, tol, from \code{\link{prcomp}}.}
}
\value{
An object of class "gm.prcomp" is a list with the following components:
\item{pc.summary}{A table summarizing the percent variation explained by each pc axis, equivalent to summary of \code{\link{prcomp}}.}
\item{pc.scores}{The set of principal component scores for all specimens.}
\item{pc.shapes}{A list with the shape coordinates of the extreme ends of all PC axes, e.g. $PC1min}
\item{sdev}{The standard deviations of the principal components (i.e., the square roots of the eigenvalues of the 
covariance/correlation matrix, as per \code{\link{prcomp}}.}
\item{rotation}{The matrix of variable loadings, as per \code{\link{prcomp}}.}
\item{anc.states}{The matrix of estimated ancestral shapes, if a phylogeny is used.}
\item{anc.pcscores}{The matrix of principal component scores for the phylogeny nodes.}
}
\description{
Function performs raw or weighted PCA on superimposed shape coordinates
}
\details{
The function performs a principal components analysis of shape variation, with the possibility
of weighing the analysis using a phylogenetic tree, or a variance-covariance matrix, 
to incorporate the expected covariance among dependent observations 
(e.g. due to spatial autocorrelation). At present combined analysis allowing the use of both 
a phylogeny and another covariance matrix is not implemented.
If a phylogeny is provided and phylo.pca = TRUE, observations are phylogenetically adjusted to 
calculate a phylogenetic PCA (Revell, 2009). 
If phylo.pca = FALSE, ancestral shapes are first estimated for the phylogeny nodes and PCA is then calculated
on the combined matrix of raw and ancestor data, producing the data necessary for a phylomorphospace plot.

PLOTTING: Contrary to previous geomorph implementations, gm.prcomp does not produce plots. 
For plotting options of gm.prcomp class objects see \code{\link{plot.gm.prcomp}} and the 
examples below.
}
\examples{
data(plethspecies) 
Y.gpa<-gpagen(plethspecies$land)    #GPA-alignment

### Raw data PCA
pleth.raw <- gm.prcomp(Y.gpa$coords)
summary(pleth.raw)

# Plotting
gps <- as.factor(c(rep("gp1", 5), rep("gp2", 4))) # Two random groups
par(mgp = c(2.5, 0.5, 0))
plot(pleth.raw, pch=22, cex = 1.5, xlab = "My PCA - axis 1", bg = gps,
    font.lab = 2, cex.lab = 2) # Modify options as desired
legend("topright", pch=22, pt.bg = unique(gps), legend = levels(gps), cex = 2)

### Phylogenetic PCA
pleth.ppca <- gm.prcomp(Y.gpa$coords, plethspecies$phy, phylo.pca = TRUE)
summary(pleth.ppca) 

# Plotting
par(mgp = c(2, 0.5, 0))
plot(pleth.ppca, phylo = TRUE, cex = 1.5, pch = 22, bg = gps, cex.lab = 2, 
     font.lab = 2, xlim = c(-0.007, 0.017),
     phylo.par = list(edge.color = "grey", edge.width = 2,
     node.bg = "black", node.pch = 22, node.cex = 0.5))
text(pleth.ppca$pc.scores, labels = labels(pleth.ppca$pc.scores)[[1]],
     pos = 2, font = 4) 
text(pleth.ppca$anc.pcscores, labels = labels(pleth.ppca$anc.pcscores)[[1]],
     adj = c(-0.1, -0.1), font = 2) 
     
### Phylomorphospace
pleth.phylomorpho <- gm.prcomp(Y.gpa$coords, plethspecies$phy)
summary(pleth.phylomorpho)

# Plotting
plot(pleth.phylomorpho, phylo = TRUE, cex = 2, pch = 22, bg = gps, 
     phylo.par = list(edge.color = "blue", edge.width = 2, edge.lty = 2,
     node.cex = 0)) # Supresses plotting of nodes
text(pleth.phylomorpho$pc.scores, labels = labels(pleth.phylomorpho$pc.scores)[[1]],
     pos = 3, font = 4) 

# same plot but change the PC axes type

plot(pleth.phylomorpho, phylo = TRUE, cex = 2, pch = 22, bg = gps, 
     phylo.par = list(edge.color = "blue", edge.width = 2, edge.lty = 2,
     node.cex = 0), axes = FALSE) # Supresses plotting of nodes and axes
 abline(h = 0, lty = 3, cex = 0.7) # subtle axes
 abline(v = 0, lty = 3, cex = 0.7)
text(pleth.phylomorpho$pc.scores, labels = labels(pleth.phylomorpho$pc.scores)[[1]],
     pos = 3, font = 4) 
}
\references{
Revell, L. J. (2009) Size-correction and principal components for interspecific comparative studies. 
Evolution, 63, 3258-3268.
}
\author{
Antigoni Kaliontzopoulou
}
\keyword{visualization}
