% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gridPar.r
\name{gridPar}
\alias{gridPar}
\title{Set up parameters for grids, points, and links in plotRefToTarget}
\usage{
gridPar(pt.bg = "black", pt.size = 1.5, n.col.cell = 20,
  grid.col = "black", grid.lwd = 1, grid.lty = 1, link.col = "black",
  link.lwd = 2, link.lty = 1, tar.pt.bg = "gray", tar.pt.size = 1,
  tar.link.col = "gray", tar.link.lwd = 2, tar.link.lty = 1)
}
\arguments{
\item{pt.bg}{Background color of reference configuration points (single value or vector of values)}

\item{pt.size}{Scale factor for reference configuration points (single value or vector of values)}

\item{n.col.cell}{The number of square cells (along x axis) for grids (single numerical value)}

\item{grid.col}{The color of grid lines (single value)}

\item{grid.lwd}{Scale factor for the weight of grid lines (single numerical value)}

\item{grid.lty}{The line type for grid lines (single numerical value, as in base R plot function)}

\item{link.col}{The color of links for reference configurations (single value or vector of values)}

\item{link.lwd}{The line weight of links for reference configurations (single value or vector of values)}

\item{link.lty}{The line type of links for reference configurations (single value or vector of values)}

\item{tar.pt.bg}{Background color of target configuration points (single value or vector of values)}

\item{tar.pt.size}{Scale factor for target configuration points (single value or vector of values)}

\item{tar.link.col}{The color of links for target configurations (single value or vector of values)}

\item{tar.link.lwd}{The line weight of links for target configurations (single value or vector of values)}

\item{tar.link.lty}{The line type of links for target configurations (single value or vector of values)}
}
\description{
Function produces a list of parameter changes within \code{\link{plotRefToTarget}} (2D only)
}
\details{
The function allows users to vary certain plotting parameters to produce different
graphical outcomes for \code{\link{plotRefToTarget}}.  Not all parameters need to be adjusted to use this
function.  This function currently only works with 2D shape data.
}
\examples{
data(plethodon)
Y.gpa<-gpagen(plethodon$land)    #GPA-alignment

ref<-mshape(Y.gpa$coords)

plotRefToTarget(ref,Y.gpa$coords[,,39]) # default settings

GP1 <- gridPar(pt.bg = "red", pt.size = 1, link.col="blue", link.lwd=2, n.col.cell=50)
plotRefToTarget(ref,Y.gpa$coords[,,39], gridPars=GP1, mag=2, links=plethodon$links, method="TPS")

GP2 <- gridPar(pt.bg = "green", pt.size = 1)
plotRefToTarget(ref,Y.gpa$coords[,,39], gridPars=GP2, mag=3, method="vector")

GP3 <- gridPar(pt.bg = "blue", pt.size = 1.5, tar.pt.bg = "orange", tar.pt.size = 1)
plotRefToTarget(ref,Y.gpa$coords[,,39], gridPars=GP3, mag=3, method="points")
}
\author{
Michael Collyer
}
\seealso{
\code{\link{plotRefToTarget}}
}
\keyword{visualization}

