% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/integration.test.R
\name{integration.test}
\alias{integration.test}
\title{Quantify morphological integration between modules}
\usage{
integration.test(A, A2 = NULL, partition.gp = NULL, iter = 999)
}
\arguments{
\item{A}{A 3D array (p x k x n) containing GPA-aligned coordinates for all specimens, or a matrix (n x variables)}

\item{A2}{An optional 3D array (p x k x n) containing GPA-aligned coordinates for all specimens, or a matrix (n x variables) for a second partition}

\item{partition.gp}{A list of which landmarks (or variables) belong in which partition (e.g. A,A,A,B,B,B,C,C,C) (required when only 1 dataset provided)}

\item{iter}{Number of iterations for significance testing}
}
\value{
Objects of class "pls" from integration.test return a list of the following:
  \item{r.pls}{The estimate of morphological integration: PLS.corr}
  \item{r.pls.mat}{The pairwise r.pls, if the number of paritions is greater than 2}
  \item{P.value}{The empirically calculated P-value from the resampling procedure.}
  \item{left.pls.vectors}{The singular vectors of the left (x) block (for 2 modules only)}
  \item{right.pls.vectors}{The singular vectors of the right (y) block (for 2 modules only)}
  \item{random.r}{The correlation coefficients found in each random permutation of the
  resampling procedure.}
  \item{XScores}{Values of left (x) block projected onto singular vectors
  (for 2 modules only).}
  \item{YScores}{Values of right (y) block projected onto singular vectors
  (for 2 modules only).}
  \item{A1}{Input values for the left block (for 2 modules only).}
  \item{A2}{Input values for the right block (for 2 modules only).}
  \item{A1.matrix}{Left block (matrix) found from A1 (for 2 modules only).}
  \item{A2.matrix}{Right block (matrix) found from A2 (for 2 modules only).}
  \item{permutations}{The number of random permutations used in the resampling procedure.}
  \item{call}{The match call.}
}
\description{
Function quantifies the degree of morphological integration between modules of Procrustes-aligned
  coordinates
}
\details{
The function quantifies the degree of morphological integration between modular partitions of shape data as
  defined by landmark coordinates. It is assumed that the landmarks have previously been aligned using
  Generalized Procrustes Analysis (GPA) [e.g., with \code{\link{gpagen}}]. The function may be used to assess
  the degree of morphological integration between two or more sets of variables.

The function estimates the degree of morphological integration using a two-block partial least squares
  analysis (PLS). When used with landmark data, this analysis is referred to as singular warps analysis
  (Bookstein et al. 2003). If more than two partitions are defined, the average pairwise PLS correlation is utilized as
  the test statistic. The observed test value is then compared to a distribution of values obtained by randomly permuting
  the individuals (rows) in one partition relative to those in the other. A significant result is found when the
  observed PLS correlation is large relative to this distribution, and implies that the structures are integrated with one
  another (see Bookstein et al. 2003).  If only two partitions are specified, a plot of PLS scores along the first
  set of PLS axes is optionally displayed, and thin-plate spline deformation grids along these axes are also shown if data were
  input as a 3D array.

Input for the analysis can take one of two forms. First, one can input a single dataset (as a matrix or 3D array, along with
 a vector describing which variables correspond to which partitions (for the case of a 3D array, which landmarks belong to which
 partitions is specified). Alternatively, when evaluating the integration between two structures or partitions, two datasets may be provided.
}
\examples{
data(plethodon)
Y.gpa<-gpagen(plethodon$land)    #GPA-alignment
#landmarks on the skull and mandible assigned to partitions
land.gps<-c("A","A","A","A","A","B","B","B","B","B","B","B")
IT <- integration.test(Y.gpa$coords, partition.gp=land.gps, iter=999)
summary(IT) # Test summary
plot(IT) # PLS plot
IT$left.pls.vectors # extracting just the left (first block) singular vectors
}
\author{
Dean Adams
}
\references{
Bookstein, F. L., P. Gunz, P. Mitteroecker, H. Prossinger, K. Schaefer, and H. Seidler.
  2003. Cranial integration in Homo: singular warps analysis of the midsagittal plane in ontogeny and
  evolution. J. Hum. Evol. 44:167-187.
}
\keyword{analysis}

