% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.defunct.r
\name{morphol.integr}
\alias{morphol.integr}
\title{Deprecated functions in geomorph}
\usage{
morphol.integr()
}
\description{
The following function has been deprecated in geomorph
}
\details{
This function has been deprecated. Below shows the original function and the replacement function.

morphol.integr now deprecated: use \code{\link{integration.test}} instead
}

