% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.defunct.r
\name{nested.update}
\alias{nested.update}
\title{Deprecated functions in geomorph}
\usage{
nested.update()
}
\description{
The following function has been deprecated in geomorph
}
\details{
This function has been deprecated. Use procD.lm, followed by \code{\link{anova.lm.rrpp}}, instead, with error adjustment.  
See examples in \code{\link{procD.lm}}
}
