% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pairwise.slope.test.r
\name{pairwise.slope.test}
\alias{pairwise.slope.test}
\title{Deprecated functions in geomorph}
\usage{
pairwise.slope.test()
}
\description{
The following function has been deprecated in geomorph
}
\details{
This function has been deprecated. Below shows the original function and the replacement function.

pairwise.slope.test: use 'advanced.procD.lm' instead
}

