% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pairwiseD.test.r
\name{pairwiseD.test}
\alias{pairwiseD.test}
\title{PairwiseD.test
#'
Function deprecated: use 'advanced.procD.lm' instead}
\usage{
pairwiseD.test()
}
\description{
PairwiseD.test
#'
Function deprecated: use 'advanced.procD.lm' instead
}

