% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pairwiseD.test.r
\name{pairwiseD.test}
\alias{pairwiseD.test}
\title{Deprecated functions in geomorph}
\usage{
pairwiseD.test()
}
\description{
The following function has been deprecated in geomorph
}
\details{
This function has been deprecated. Below shows the original function and the replacement function.

pairwiseD.test: use 'advanced.procD.lm' instead
}

