% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylo.integration.R
\name{phylo.integration}
\alias{phylo.integration}
\title{Quantify phylogenetic morphological integration between two or more sets of variables under Brownian motion}
\usage{
phylo.integration(A, A2 = NULL, phy, partition.gp = NULL, iter = 999)
}
\arguments{
\item{A}{A 2D array (n x [p1 x k1]) or 3D array (p1 x k1 x n) containing landmark coordinates for the first block}

\item{A2}{An optional 2D array (n x [p2 x k2]) or 3D array (p2 x k2 x n) containing landmark coordinates for the second block}

\item{phy}{A phylogenetic tree of {class phylo} - see \code{\link[ape]{read.tree}} in library ape}

\item{partition.gp}{A list of which landmarks (or variables) belong in which partition (e.g. A,A,A,B,B,B,C,C,C) (required when only 1 dataset provided)}

\item{iter}{Number of iterations for significance testing}
}
\value{
Function returns a list with the following components:
  \item{PLS.corr}{The estimate of phylogenetic morphological covariation}
  \item{pvalue}{The significance level of the observed signal}
  \item{Xscores}{PLS scores for the first block of landmarks (for 2 modules only)}
  \item{Yscores}{PLS scores for the second block of landmarks (for 2 modules only)}
}
\description{
Function quantifies the degree of phylogenetic morphological covariation between two or more sets of
Procrustes-aligned coordinates using partial least squares.
}
\details{
The function quantifies the degree of phylogenetic morphological integration between two or more sets of shape data as
  defined by landmark coordinates. The approach is based on a Brownian motion model of evolution. It is
  assumed that the landmarks have previously been aligned using
  Generalized Procrustes Analysis (GPA) [e.g., with \code{\link{gpagen}}].

The function estimates the degree of  morphological covariation between two or sets of variables
while accounting for phylogeny using partial least squares (Adams and Felice 2014), and under a Brownian
motion model of evolution. If more than two partitions
are defined, the average pairwise PLS correlation is utilized as the test statistic. The observed value is
statistically assessed using permutation, where data for one partition are permuted across the tips of the phylogeny,
an estimate of the covariation between sets of variables is obtained, and compared to the observed value. The analysis is conducted
under a Brownian motion model of evolution (see Adams anbd Felice 2014).

Input for the analysis can take one of two forms. First, one can input a single dataset (as a matrix or 3D array, along with
 a vector describing which variables correspond to which partitions (for the case of a 3D array, which landmarks belong to which
 partitions is specified). Alternatively, when evaluating the integration between two structures or partitions, two datasets may be provided.

The generic functions, \code{\link{print}}, \code{\link{summary}}, and \code{\link{plot}} all work with \code{\link{phylo.integration}}.
 The generic function, \code{\link{plot}}, produces a two-block.pls plot.  This function calls \code{\link{plot.pls}}, which has two additional
 arguments (with defaults): label = NULL, warpgrids = TRUE.  These arguments allow one to include a vector to label points and a logical statement to
 include warpgrids, respectively.  Warpgrids can only be included for 3D arrays of Procrustes resdiuals. The plot is a plot of PLS scores from
 Block1 versus Block2 performed for the first set of PLS axes.
}
\examples{
data(plethspecies)
Y.gpa<-gpagen(plethspecies$land)    #GPA-alignment
land.gps<-c("A","A","A","A","A","B","B","B","B","B","B")

IT<- phylo.integration(Y.gpa$coords,partition.gp=land.gps,phy=plethspecies$phy,iter=999)
summary(IT) # Test summary
plot(IT) # PLS plot
}
\author{
Dean Adams
}
\references{
Adams, D.C. and R. Felice. 2014. Assessing phylogenetic morphological
integration and trait covariation in morphometric data using evolutionary covariance
matrices. PLOS ONE. 9(4):e94335.
}
\seealso{
\code{\link{integration.test}}, \code{\link{modularity.test}}, \code{\link{phylo.pls}}, and
\code{\link{two.b.pls}}
}
\keyword{analysis}

