% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.defunct.r
\name{phylo.pls}
\alias{phylo.pls}
\title{Deprecated functions in geomorph}
\usage{
phylo.pls()
}
\description{
The following function has been deprecated in geomorph
}
\details{
This function has been deprecated. Below shows the original function and the replacement function.

phylo.pls now deprecated: use \code{\link{phylo.integration}} instead
}

