% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picknplot.shape.r
\name{picknplot.shape}
\alias{picknplot.shape}
\title{Pick points in geomorph scatterplots to visualize shape variation}
\usage{
picknplot.shape(x, ...)
}
\arguments{
\item{x}{a geomorph plot object of class plot.gm.prcomp, plot.procD.lm, plot.trajectory.analysis or plot.pls}

\item{...}{other arguments passed to \code{\link{plotRefToTarget}}}
}
\value{
A list with the following components:
\item{points}{A list with the xy coordinates of the selected points.}
\item{shapes}{A list with the corresponding estimated shapes.}
}
\description{
Function plots the shape corresponding to a clicked point in the area of a geomorph plot
}
\details{
This function makes the scatterplots created by generic \code{\link{plot}} functions for geomorph analytical objects interactive, 
which allows one to visualize shape variation by selecting one or more points in morphospace. The function uses \code{\link{shape.predictor}} 
to estimate the shape corresponding to the selected point(s) based on the prediction underlying the scatterplot, and it plots the estimated 
shape as compared to the consensus landmark configuration using \code{\link{plotRefToTarget}}. The user is then prompted as to whether the plotted
shape is to be saved as a png file, in which case the name of the file needs to be provided (without quotation marks).
Interactive plots are at present available for plots produced by \code{\link{plot.gm.prcomp}}.
}
\examples{
# 2d
data(plethodon) 
Y.gpa <- gpagen(plethodon$land)
pleth.pca <- gm.prcomp(Y.gpa$coords)
pleth.pca.plot <- plot(pleth.pca$rawPCA)
picknplot.shape(pleth.pca.plot) 
# May change arguments for plotRefToTarget
picknplot.shape(plot(pleth.pca$rawPCA), method = "points", mag = 3, links=plethodon$links)

# 2d with phylogeny
data(plethspecies) 
Y.gpa <- gpagen(plethspecies$land)
gps <- as.factor(c(rep("gp1", 5), rep("gp2", 4))) # Two random groups
pleth.phylo <- gm.prcomp(Y.gpa$coords, plethspecies$phy)
pleth.phylomorphospace <- plot(pleth.phylo$phylomorphospace, phylo = TRUE, cex = 2, pch = 22, 
bg = gps, phylo.par = list(edge.color = "blue", edge.width = 2, edge.lty = 2,
node.pch = 22, node.bg = "black"))
links.species <- plethodon$links[-11,]
links.species[11, 1] <- 11
picknplot.shape(pleth.phylomorphospace, method = "points", links = links.species)
# 3d
data(scallops)
Y3d.gpa <- gpagen(A=scallops$coorddata, curves=scallops$curvslide, surfaces=scallops$surfslide)
scallops.pca <- gm.prcomp(Y3d.gpa$coords)
scallops.pca.plot <- plot(scallops.pca$rawPCA)
picknplot.shape(scallops.pca.plot) #May change method for plotRefToTarget
}
\seealso{
\code{\link{shape.predictor}}, \code{\link{plotRefToTarget}}

\code{\link[rgl]{rgl-package}} (used in 3D plotting)
}
\author{
Antigoni Kaliontzopoulou & Emma Sherratt
}
\keyword{visualization}
