% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{plot.advanced.procD.lm}
\alias{plot.advanced.procD.lm}
\title{Plot Function for geomorph}
\usage{
\method{plot}{advanced.procD.lm}(x, ...)
}
\arguments{
\item{x}{plot object (from \code{\link{advanced.procD.lm}})}

\item{...}{other arguments passed to  \code{\link{plot.procD.lm}}}
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}
