% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.R
\name{plot.bilat.symmetry}
\alias{plot.bilat.symmetry}
\title{Plot Function for geomorph}
\usage{
\method{plot}{bilat.symmetry}(x, warpgrids = TRUE, mesh = NULL, ...)
}
\arguments{
\item{x}{plot object}

\item{warpgrids}{Logical argument whether to include warpgrids}

\item{mesh}{Option to include mesh in warpgrids plots}

\item{...}{other arguments passed to plot}
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}

