% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{plot.CR.phylo}
\alias{plot.CR.phylo}
\title{Plot Function for geomorph}
\usage{
\method{plot}{CR.phylo}(x, ...)
}
\arguments{
\item{x}{plot object (from \code{\link{phylo.modularity}})}

\item{...}{other arguments passed to plot}
}
\description{
Plot Function for geomorph
}
\author{
Dean Adams
}
\keyword{utilities}
\keyword{visualization}
