% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{plot.gm.prcomp}
\alias{plot.gm.prcomp}
\title{Plot Function for geomorph}
\usage{
\method{plot}{gm.prcomp}(x, axis1 = 1, axis2 = 2, phylo = FALSE,
  phylo.par = list(edge.color = "black", edge.width = 1, edge.lty = 1,
  node.bg = "black", node.pch = 21, node.cex = 1), ...)
}
\arguments{
\item{x}{An object of class \code{\link{gm.prcomp}}}

\item{axis1}{A value indicating which PC axis should be displayed as the X-axis (default = PC1)}

\item{axis2}{A value indicating which PC axis should be displayed as the Y-axis (default = PC2)}

\item{phylo}{A logical value indicating whether the phylogeny should be projected to PC space}

\item{phylo.par}{A list of plotting parameters for the phylogeny edges (edge.color, edge.width, edge.lty)
and nodes (node.bg, node.pch, node.cex)}

\item{...}{other arguments passed to plot}
}
\value{
An object of class "plot.gm.prcomp" is a list with components
 that can be used in other plot functions, such as the type of plot, points, 
 a group factor, and other information depending on the plot parameters used.
 
 NOTE: To visualize shape variation across PC axes, use \code{\link{picknplot.shape}}.
}
\description{
Plot Function for geomorph
}
\seealso{
\code{\link{plotRefToTarget}}
}
\author{
Antigoni Kaliontzopoulou
}
\keyword{utilities}
\keyword{visualization}
