% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{plot.gpagen}
\alias{plot.gpagen}
\title{Plot Function for geomorph}
\usage{
\method{plot}{gpagen}(x, ...)
}
\arguments{
\item{x}{plot object (from \code{\link{gpagen}})}

\item{...}{other arguments passed to plotAllSpecimens}
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}
