% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{plot.module.eigen}
\alias{plot.module.eigen}
\title{Plot function for geomorph}
\usage{
\method{plot}{module.eigen}(x, use.rel.dims = TRUE, add.lines = TRUE, add.legend = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{\link{module.eigen}}}

\item{use.rel.dims}{A logical argument for whether to use only the relevant dimensions in the plot, 
based on a broken stick model.}

\item{...}{other arguments passed to plot.

This function vectorizes a summary table (for proportions of variance) 
from \code{\link{summary.module.eigen}}.  For example, if a 5 x 4 table is produced, a vector of length 20 is produced
by the columns of the table.  Some plot parameters can be manipulated, but one has to do it with respect to all values.  
For example, for the 5 x 4 table, bg = c(rep(1:4, each = 5)) produces a vector of length 20 to describe point 
background colors.}
}
\description{
Plot function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
