% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.R
\name{plot.physignal}
\alias{plot.physignal}
\title{Plot Function for geomorph}
\usage{
\method{plot}{physignal}(x, ...)
}
\arguments{
\item{x}{plot object}

\item{...}{other arguments passed to plot}
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}

