% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.R
\name{plot.pls}
\alias{plot.pls}
\title{Plot Function for geomorph}
\usage{
\method{plot}{pls}(x, label = NULL, warpgrids = TRUE, ...)
}
\arguments{
\item{x}{plot object}

\item{label}{Optional vector to label points}

\item{warpgrids}{Logical argument whether to include warpgrids}

\item{...}{other arguments passed to plot}
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}

