% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.R
\name{plot.procD.allometry}
\alias{plot.procD.allometry}
\title{Plot Function for geomorph}
\usage{
\method{plot}{procD.allometry}(x, method = c("CAC", "RegScore", "PredLine"),
  warpgrids = TRUE, label = NULL, mesh = NULL, ...)
}
\arguments{
\item{x}{plot object}

\item{method}{Method for estimating allometric shape components; see \code{\link{procD.allometry}} for details}

\item{warpgrids}{A logical value indicating whether deformation grids for small and large shapes 
should be displayed (note: if groups are provided no TPS grids are shown)}

\item{label}{An optional vector indicating labels for each specimen that are to be displayed}

\item{mesh}{A mesh3d object to be warped to represent shape deformation of the minimum and maximum size 
if {warpgrids=TRUE} (see \code{\link{warpRefMesh}}).}

\item{...}{other arguments passed to plot}
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}

