% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.R
\name{plot.procD.lm}
\alias{plot.procD.lm}
\title{Plot Function for geomorph}
\usage{
\method{plot}{procD.lm}(x, outliers = FALSE, ...)
}
\arguments{
\item{x}{plot object}

\item{outliers}{Logical argument to include outliers plot}

\item{...}{other arguments passed to plot}
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}

