% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.R
\name{plot.trajectory.analysis}
\alias{plot.trajectory.analysis}
\title{Plot Function for geomorph}
\usage{
\method{plot}{trajectory.analysis}(x, group.cols = NULL,
  pt.seq.pattern = c("white", "gray", "black"), pt.scale = 1, ...)
}
\arguments{
\item{x}{plot object}

\item{group.cols}{An optional vector of colors for group levels}

\item{pt.seq.pattern}{The sequence of colors for starting, middle, and end points of 
trajectories, respectivly.  E.g., c("green", "gray", "red") for gray points
but initial points with green color and end points with red color.}

\item{pt.scale}{An optional value to magnify or reduce points (1 = no change)}

\item{...}{other arguments passed to plot}
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}

