% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{plot.trajectory.analysis}
\alias{plot.trajectory.analysis}
\title{Plot Function for geomorph}
\usage{
\method{plot}{trajectory.analysis}(x, group.cols = NULL,
  pt.seq.pattern = c("white", "gray", "black"), pt.scale = 1, ...)
}
\arguments{
\item{x}{plot object (from \code{\link{trajectory.analysis}})}

\item{group.cols}{An optional vector of colors for group levels}

\item{pt.seq.pattern}{The sequence of colors for starting, middle, and end points of 
trajectories, respectively.  E.g., c("green", "gray", "red") for gray points
but initial points with green color and end points with red color.}

\item{pt.scale}{An optional value to magnify or reduce points (1 = no change)}

\item{...}{other arguments passed to plot}
}
\value{
An object of class "plot.trajectory.analysis" is a list with components
 that can be used in other plot functions, such as the type of plot, points, 
 a group factor, and other information depending on the plot parameters used.
}
\description{
Plot Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}
