% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.defunct.r
\name{plotAllometry}
\alias{plotAllometry}
\title{Deprecated functions in geomorph}
\usage{
plotAllometry()
}
\description{
The following function has been deprecated in geomorph
}
\details{
This function has been deprecated. Below shows the original function and the replacement function.

plotAllometry now deprecated: use \code{\link{procD.allometry}} instead
}

