% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotAllometry.R
\name{plotAllometry}
\alias{plotAllometry}
\title{Plot allometric patterns in landmark data}
\usage{
plotAllometry(A, sz, groups = NULL, method = c("CAC", "RegScore",
  "PredLine"), warpgrids = TRUE, iter = 249, label = NULL, mesh = NULL,
  logsz = TRUE, verbose = FALSE)
}
\arguments{
\item{A}{matrix (n x [p1 x k]) or 3D array (p1 x k x n) containing GPA-aligned coordinates for the specimens}

\item{sz}{A vector of size measures for all specimens (centroid size or any other reasonable size measure)}

\item{groups}{An optional vector containing group labels for each specimen if available}

\item{method}{Method for estimating allometric shape components; see below for details}

\item{warpgrids}{A logical value indicating whether deformation grids for small and large shapes
should be displayed (note: if groups are provided no TPS grids are shown)}

\item{iter}{Number of iterations for significance testing}

\item{label}{An optional vector indicating labels for each specimen that are to be displayed}

\item{mesh}{A mesh3d object to be warped to represent shape deformation of the directional and fluctuating components
of asymmetry if {warpgrids= TRUE} (see \code{\link{warpRefMesh}}).}

\item{logsz}{A logical value indicating whether log(size) is used}

\item{verbose}{A logical value indicating whether the output is basic or verbose (see Value below)}
}
\value{
Function returns an ANOVA table of statistical results for size: df, SS, MS, Prand.
If verbose=TRUE, function returns a list with the following components:
 \item{ProcDist.lm}{An ANOVA table as above}
 \item{allom.score}{ A matrix of the allometry shape scores}
 \item{logSize}{ A matrix of log size}
 \item{pred.shape}{A matrix containing the predicted shapes from the regression}
 \item{resid.shape}{ The residual shape component (RSC) of the data ("method=CAC" only)}
}
\description{
Function plots allometry curves for a set of specimens
}
\details{
The function performs a regression of shape on size, and generates a plot that describes the
multivariate relationship between size and shape
  derived from landmark data (i.e., allometry). It is assumed that the landmarks have previously been
  aligned using Generalized Procrustes Analysis (GPA) [e.g., with \code{\link{gpagen}}]. The abscissa
  of the plot is log(centroid size) while the ordinate represents shape [NOTE: the function takes the
  input size and performed log-transformation automatically by default (logsz = TRUE), as log(Size) should be used].
  Three complementary approaches can be implemented to visualize allometry:
 \enumerate{
  \item {If "method=CAC" (the default) the function calculates the
  common allometric component of the shape data, which is an estimate of the average allometric trend
  within groups (Mitteroecker et al. 2004). The function also calculates the residual shape component (RSC) for
  the data.}
  \item {If "method=RegScore" the function calculates shape scores
  from the regression of shape on size, and plots these versus size (Drake and Klingenberg 2008).
  For a single group, these shape scores are mathematically identical to the CAC (Adams et al. 2013).}
  \item {If "method=PredLine" the function calculates predicted values from a regression of shape on size, and
  plots the first principal component of the predicted values versus size as a stylized graphic of the
  allometric trend (Adams and Nistri 2010). }
  }
  For all methods, both centroid size and allometry scores are returned. Optionally, deformation grids can be
  requested, which display the shape of the smallest and largest specimens relative to the average specimen (using
  'warpgrids=T' or 'warpgrids=F').
  Finally, if groups are provided, the above approaches are implemented while
  accounting for within-group patterns of covariation (see references for explanation). In this case,
  the regression is of the form: shape~size+groups (Note: to examine the interaction term use \code{\link{procD.lm}}).
  Specimens from each group are plotted using distinct colors based on the order in which the groups are
  found in the dataset, and using R's standard color palette: black, red, green, blue, cyan, magenta,
  yellow, and gray. NOTE: to change the colors of the groups, simply substitute a vector of the desired colors for
  each specimen.
}
\examples{
data(ratland)
Y.gpa<-gpagen(ratland,PrinAxes=FALSE)    #GPA-alignment

#Using CAC for plot
plotAllometry(Y.gpa$coords,Y.gpa$Csize,method="CAC", iter=5)

#Using Regression Scores for plot
plotAllometry(Y.gpa$coords,Y.gpa$Csize,method="RegScore", iter=5)

#Using predicted allometry curve for plot
plotAllometry(Y.gpa$coords,Y.gpa$Csize,method="PredLine", iter=5)
}
\author{
Dean Adams
}
\references{
Adams, D.C., F.J. Rohlf, and D.E. Slice. 2013. A field comes of age: geometric morphometrics
  in the 21st century. Hystrix. 24:7-14.

Adams, D. C., and A. Nistri. 2010. Ontogenetic convergence and evolution of foot morphology
  in European cave salamanders (Family: Plethodontidae). BMC Evol. Biol. 10:1-10.

Drake, A. G., and C. P. Klingenberg. 2008. The pace of morphological change: Historical
  transformation of skull shape in St Bernard dogs. Proceedings of the Royal Society B, Biological Sciences 275:71'76.

Mitteroecker, P., P. Gunz, M. Bernhard, K. Schaefer, and F. L. Bookstein. 2004.
  Comparison of cranial ontogenetic trajectories among great apes and humans. J. Hum. Evol. 46:679-698.
}
\keyword{analysis}
\keyword{visualization}

