% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGMPhyloMorphoSpace.r
\name{plotGMPhyloMorphoSpace}
\alias{plotGMPhyloMorphoSpace}
\title{Plot phylogenetic tree and specimens in tangent space}
\usage{
plotGMPhyloMorphoSpace()
}
\description{
This function is now defunct
}
