% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.defunct.r,
%   R/plotGMPhyloMorphoSpace.r
\name{plotGMPhyloMorphoSpace}
\alias{plotGMPhyloMorphoSpace}
\title{Deprecated and defunct functions in geomorph}
\usage{
plotGMPhyloMorphoSpace()

plotGMPhyloMorphoSpace()
}
\description{
The following function is now defunct

This function is now defunct
}
