% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.defunct.r, R/plotTangentSpace.r
\name{plotTangentSpace}
\alias{plotTangentSpace}
\title{Deprecated and defunct functions in geomorph}
\usage{
plotTangentSpace()

plotTangentSpace()
}
\description{
The following function is now defunct

This function is now defunct
}
