% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{print.CR}
\alias{print.CR}
\title{Print/Summary Function for geomorph}
\usage{
\method{print}{CR}(x, ...)
}
\arguments{
\item{x}{print/summary object (from \code{\link{phylo.modularity}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
