% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{print.trajectory.analysis}
\alias{print.trajectory.analysis}
\title{Print/Summary Function for geomorph}
\usage{
\method{print}{trajectory.analysis}(x, angle.type = c("r", "rad", "deg"), ...)
}
\arguments{
\item{x}{print/summary object}

\item{angle.type}{Choice between vector correlation or vector angles, in radians or degrees
for summarizing results ("r", "rad", "deg", respectively)}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
