% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ratland}
\alias{ratland}
\title{Landmark data from dataset rat}
\description{
Landmark data from dataset rat
}
\author{
Dean Adams
}
\references{
Bookstein, F. L. 1991. Morphometric tools for landmark data: Geometry and Biology.
 Cambridge Univ. Press, New York.
}
\keyword{datasets}

