% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readland.fcsv.r
\name{readland.fcsv}
\alias{readland.fcsv}
\title{Read landmark data matrix from fcsv file}
\usage{
readland.fcsv(file = NULL)
}
\arguments{
\item{file}{the name of a *.fcsv file containing three-dimensional landmark data to be read in}
}
\value{
Function returns a p x 3 matrix of x, y, z coordinates for p landmarks.
}
\description{
Read landmark data for a single specimen from an fcsv file obtained from SlicerMorph.
}
\details{
This function merely extracts x, y, z coordinates from an fcsv file.
}
\author{
Murat Maga and Michael Collyer
}
\keyword{IO}
