% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.support.code.r
\docType{data}
\name{scallopPLY}
\alias{scallopPLY}
\title{3D scan of a scallop shell from a .ply file in mesh3d format}
\description{
3D scan of a scallop shell from a .ply file in mesh3d format
}
\author{
Emma Sherratt
}
\references{
Serb et al. (2011). "Morphological convergence of shell shape in distantly related
scallop species (Mollusca: Pectinidae)." Zoological Journal of the Linnean Society 163: 571-584.
}
\keyword{datasets}

