\docType{data}
\name{scallops}
\alias{scallops}
\title{Landmark data from scallop shells}
\description{
  Landmark data from scallop shells
}
\author{
  Dean Adams and Erik Otarola-Castillo
}
\references{
  Serb et al. (2011). "Morphological convergence of shell
  shape in distantly related scallop species (Mollusca:
  Pectinidae)." Zoological Journal of the Linnean Society
  163: 571-584.
}
\keyword{data}
\keyword{datasets}

