% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{summary.bilat.symmetry}
\alias{summary.bilat.symmetry}
\title{Print/Summary Function for geomorph}
\usage{
\method{summary}{bilat.symmetry}(object, ...)
}
\arguments{
\item{object}{print/summary object (from \code{\link{bilat.symmetry}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
