% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{summary.CR.phylo}
\alias{summary.CR.phylo}
\title{Print/Summary Function for geomorph}
\usage{
\method{summary}{CR.phylo}(object, ...)
}
\arguments{
\item{object}{print/summary object (from \code{\link{phylo.modularity}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Dean Adams
}
\keyword{utilities}
