% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.R
\name{summary.CR}
\alias{summary.CR}
\title{Print/Summary Function for geomorph}
\usage{
\method{summary}{CR}(object, ...)
}
\arguments{
\item{object}{print/summary object}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}

