% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{summary.module.eigen}
\alias{summary.module.eigen}
\title{Print/Summary function for geomorph}
\usage{
\method{summary}{module.eigen}(object, use.rel.dims = TRUE, PC = 1)
}
\arguments{
\item{object}{An object of class \code{\link{module.eigen}}}

\item{use.rel.dims}{A logical argument for whether to use only the relevant dimensions in the summary, 
based on a broken stick model.}

\item{PC}{A numeric value to indicate for which principal component to perform vector correlations}

\item{...}{other arguments passed to print/summary}
}
\value{
An object of class "summary.module.eigen" is a list with components
including table of eigenvalues, table of proportion of variation explained,
a pairwise table of vector correlations (for indicated principal component), and a apirwise
table of Krzanowski squared correlations.
}
\description{
Print/Summary function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
