% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{summary.morphol.disparity}
\alias{summary.morphol.disparity}
\title{Print/Summary Function for geomorph}
\usage{
\method{summary}{morphol.disparity}(object, ...)
}
\arguments{
\item{object}{print/summary object (from \code{\link{morphol.disparity}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
