% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorph.utils.r
\name{summary.pls}
\alias{summary.pls}
\title{Print/Summary Function for geomorph}
\usage{
\method{summary}{pls}(object, ...)
}
\arguments{
\item{object}{print/summary object (from \code{\link{phylo.integration}} or \code{\link{two.b.pls}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for geomorph
}
\author{
Michael Collyer
}
\keyword{utilities}
